0.4.0开发-工作流相关接口调整

+++ rosettaflow-mapper/src/main/java/com/moirae/rosettaflow/mapper/domain/WorkflowSettingExpert.java
@@ -38,7 +38,7 @@ public class WorkflowSettingExpert implements Serializable {
      * 工作流版本号
      */
     @TableField("workflow_version")
-    private Integer workflowVersion;
+    private Long workflowVersion;
 
     /**
      * 工作流中节点的顺序,从1开始
